rename default-tasks to default_tasks
