+++ discojs/discojs-core/src/default_tasks/cifar10.ts
