+++ discojs/discojs-core/src/default_tasks/mnist.ts
