+++ discojs/discojs-core/src/index.node.ts
@@ -18,6 +18,6 @@ export { Disco, TrainingSchemes } from './training'
 export { Validator } from './validation'
 
 export { TrainingInformation, DisplayInformation, isTask, Task, isTaskID, TaskID, TaskProvider, isTaskProvider, Digest, isDigest } from './task'
-export * as defaultTasks from './default-tasks'
+export * as defaultTasks from './default_tasks'
 
 export * from './types'
